split manager into extensions for readability/maintainability
