+++ .gitignore
@@ -1,4 +1,2 @@
 # ignore CLAUDE
 CLAUDE.md
-
-
